new template to split arb deployment and new tokens
