--- v3/tasks/20250429-v3-quantamm/input.ts
@@ -1,185 +1,0 @@
-import { MONTH } from '@helpers/time';
-import { Task, TaskMode } from '@src';
-import { bn, fp } from '@helpers/numbers';
-import { ZERO_ADDRESS } from '@helpers/constants';
-import { BigNumber } from 'ethers';
-
-export type QuantAMMDeploymentInputParams = {
-  Vault: string;
-  PauseWindowDuration: number;
-  UpdateWeightRunner: string;
-  ETH: string;
-  WBTC: string;
-  USDC: string;
-  FactoryVersion: string;
-  PoolVersion: string;
-  ChainlinkFeedETH: string;
-  ChainlinkDataFeedBTC: string;
-  ChainlinkDataFeedUSDC: string;
-};
-
-//TODO double check with Jeff this is network specific
-const Vault = new Task('20241204-v3-vault', TaskMode.READ_ONLY);
-
-const EthChainlinkOracleWrapper = new Task('20250419-v3-eth-oraclewrapper', TaskMode.READ_ONLY);
-
-const BtcChainlinkOracleWrapper = new Task('20250419-v3-btc-oraclewrapper', TaskMode.READ_ONLY);
-
-const UsdcChainlinkOracleWrapper = new Task('20250419-v3-usdc-oraclewrapper', TaskMode.READ_ONLY);
-
-const UpdateWeightRunner = new Task('20250419-v3-update-weight-runner', TaskMode.READ_ONLY);
-
-const BaseVersion = { version: 1, deployment: '20250429-v3-quantamm' };
-
-export default {
-  Vault,
-  ChainlinkFeedETH: EthChainlinkOracleWrapper,
-  ChainlinkDataFeedBTC: BtcChainlinkOracleWrapper,
-  ChainlinkDataFeedUSDC: UsdcChainlinkOracleWrapper,
-  PauseWindowDuration: 4 * 12 * MONTH,
-  FactoryVersion: JSON.stringify({ name: 'QuantAMMWeightedPoolFactory', ...BaseVersion }),
-  PoolVersion: JSON.stringify({ name: 'QuantAMMWeightedPool', ...BaseVersion }),
-  UpdateWeightRunner,
-  sepolia: {
-    WBTC: '0x29f2D40B0605204364af54EC677bD022dA425d03',
-    USDC: '0x94a9D9AC8a22534E3FaCa9F4e7F2E2cf85d5E4C8',
-  },
-  mainnet: {
-    WBTC: '0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599', //https://etherscan.io/token/0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599
-    USDC: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48', //https://etherscan.io/token/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48
-  },
-};
-
-type PoolRoleAccounts = {
-  // Define the structure based on Solidity contract
-};
-
-type TokenConfig = {
-  token: string;
-  rateProvider?: string;
-  tokenType: number;
-};
-
-type PoolSettings = {
-  assets: string[];
-  rule: string;
-  oracles: string[][];
-  updateInterval: number;
-  lambda: BigNumber[];
-  epsilonMax: BigNumber;
-  absoluteWeightGuardRail: BigNumber;
-  maxTradeSizeRatio: BigNumber;
-  ruleParameters: BigNumber[][];
-  poolManager: string;
-};
-
-export type CreationNewPoolParams = {
-  name: string;
-  symbol: string;
-  tokens: TokenConfig[];
-  normalizedWeights: BigNumber[];
-  roleAccounts: PoolRoleAccounts;
-  swapFeePercentage: BigNumber;
-  poolHooksContract: string;
-  enableDonation: boolean;
-  disableUnbalancedLiquidity: boolean;
-  salt: string;
-  _initialWeights: BigNumber[];
-  _poolSettings: PoolSettings;
-  _initialMovingAverages: BigNumber[];
-  _initialIntermediateValues: BigNumber[];
-  _oracleStalenessThreshold: BigNumber;
-  poolRegistry: BigNumber;
-  poolDetails: string[][];
-};
-
-export async function createPoolParams(
-  wbtcContract: string,
-  wbtcOracle: string,
-  usdcContract: string,
-  usdcOracle: string,
-  ruleAddress: string,
-  salt: string,
-  sender: string
-): Promise<CreationNewPoolParams> {
-  const tokens = [wbtcContract, usdcContract]; //address ordering as in InputHelper.sortTokens
-
-  const rateProviders: string[] = [];
-
-  const tokenConfig: TokenConfig[] = tokens.map((token, i) => ({
-    token,
-    rateProvider: rateProviders[i] || ZERO_ADDRESS,
-    tokenType: 0,
-  }));
-  //NOTE: this is order BTC, USDC
-  const lambdas = [bn('811035769801363300'), bn('289524066401247700')];
-  //const lambdas = [bn('0.811035769801363300'), bn('0.289524066401247700')];
-
-  const movingAverages = [bn('90950694357815940000000'), bn('999975466198190900')];
-  //const movingAverages = [bn('90950.69435781594'), bn('0.9999754661981909')];
-
-  const intermediateValues = [bn('101053147538869670000000'), bn('76674211432404')];
-  //const intermediateValues = [bn('101053.14753886967'), bn('0.000076674211432404264')];
-
-  //NOTE: this is order BTC, USDC
-  const parameters = [
-    [bn('1390968414526753800000'), bn('255928993330991830000')], //kappa
-    [bn('1531232793117663900'), bn('1000000000000000100')], //exponents
-  ];
-
-  //const parameters = [
-  //  [bn('1390.968414526753800000'), bn('806.695362159777100000'), bn('255.928993330991830000')], //kappa
-  //  [bn('1.531232793117663900'), bn('1.000000000000000100'), bn('1.000000000000000100')], //exponents
-  //];
-
-  //again this is in InputHelper.sortTokens order
-  const oracles = [
-    [wbtcOracle], // WBTC
-    [usdcOracle], // USDC
-  ];
-
-  const normalizedWeights = [fp(0.03), fp(0.97)];
-  const intNormalizedWeights = [...normalizedWeights];
-
-  const poolDetails = [
-    ['overview', 'adaptabilityScore', 'number', '5'],
-    ['strategy', 'name', 'string', 'Power Channel'],
-  ];
-
-  const poolSettings: PoolSettings = {
-    assets: tokens,
-    rule: ruleAddress,
-    oracles,
-    updateInterval: 86100,
-    lambda: lambdas,
-    epsilonMax: fp(0.432),
-    absoluteWeightGuardRail: fp(0.03),
-    maxTradeSizeRatio: fp(0.1),
-    ruleParameters: parameters,
-    poolManager: sender,
-  };
-
-  return {
-    name: 'TEST - DO NOT USE',
-    symbol: 'BTF:TEST',
-    tokens: tokenConfig,
-    normalizedWeights,
-    roleAccounts: {
-      pauseManager: ZERO_ADDRESS,
-      swapFeeManager: ZERO_ADDRESS,
-      poolCreator: ZERO_ADDRESS,
-    },
-    swapFeePercentage: fp(0.01),
-    poolHooksContract: ZERO_ADDRESS,
-    enableDonation: false,
-    disableUnbalancedLiquidity: false,
-    salt: salt,
-    _initialWeights: intNormalizedWeights,
-    _poolSettings: poolSettings,
-    _initialMovingAverages: movingAverages,
-    _initialIntermediateValues: intermediateValues,
-    _oracleStalenessThreshold: bn('86760'), //1 day and 1 hour
-    poolRegistry: bn('17'), //1 perform update, 3 getdata, 16 admin controlled.
-    poolDetails,
-  };
-}
